Refatoração para melhorar a legibilidade e performance do código
