+++ Anagrams/cyber-dojo-2023-2-9-4rgksR/files/test/cyberdojo/anagrams/cyberdojo2023294rgksR/files/anagramgenerator/AnagramsTest.java
@@ -1,0 +1,42 @@
+package test.cyberdojo.anagrams.cyberdojo2023294rgksR.files;
+
+import anagramgenerator.AnagramGenerator;
+import anagramgenerator.AnagramGeneratorImp;
+import anagramgenerator.Anagrams;
+import mocks.MockDisplay;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import swapper.SwapperImp;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.function.Consumer;
+
+public class AnagramsTest {
+    private Anagrams anagrams;
+    private MockDisplay display;
+    private AnagramGenerator generator;
+    private String word;
+    @BeforeEach
+    public void setup() {
+        word = "biro";
+        display = new MockDisplay();
+        generator = new AnagramGeneratorImp(new SwapperImp());
+        anagrams = new Anagrams(generator, display);
+    }
+
+    @Test
+    void should_display_all_anagrams_of_generator() {
+        List<String> anagramsList = new ArrayList<>();
+        Consumer<String> receiver = anagramsList::add;
+        generator.generateAnagrams(word, receiver);
+        anagrams.showAnagrams(word);
+        Assertions.assertEquals(anagramsList.size(), display.getCallCount("show"));
+    }
+    @Test
+    void should_display_all_anagrams() {
+        anagrams = new Anagrams(generator, System.out::println);
+        anagrams.showAnagrams(word);
+    }
+}
