+++ Anagrams/cyber-dojo-2023-2-9-4rgksR/files/test/cyberdojo/anagrams/cyberdojo2023294rgksR/files/swapper/SwapperImpTest.java
@@ -1,7 +1,9 @@
-package swapper;
+package test.swapper;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import swapper.Swapper;
+import swapper.SwapperImp;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
